requirements.txt
fastapi
uvicorn
tensorflow
mtcnn
numpy
opencv-python
pillow
python-multipart